feat: change migrations library from ley to node-pg-migrate
