+++ package.json
@@ -11,9 +11,7 @@
     "testenv:run": "docker-compose -f docker/docker-compose.dev.postgres.yml up",
     "testenv:stop": "docker-compose -f docker/docker-compose.dev.postgres.yml down -v -t 0",
     "testenv:logs": "docker-compose -f docker/docker-compose.dev.postgres.yml logs -t -f",
-    "migrate": "ley -r ts-node/register",
-    "db-down": "ley -r ts-node/register down --all",
-    "db-reset": "ley -r ts-node/register down --all && ley -r ts-node/register up",
+    "migrate": "ts-node node_modules/.bin/node-pg-migrate -j ts",
     "lint:eslint": "eslint . --ext .js,.jsx,.ts,.tsx -f unix",
     "lint:prettier": "prettier --check src/**/*.ts tests/**/*.ts migrations/**/*.ts",
     "generate:openapi": "node -r ts-node/register ./util/openapi-generator.ts",
@@ -35,7 +33,6 @@
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-tsdoc": "^0.2.17",
     "jest": "^28.1.3",
-    "ley": "^0.6.0",
     "prettier": "^2.7.1",
     "redoc-cli": "^0.13.20",
     "rimraf": "^3.0.2",
@@ -54,6 +51,7 @@
     "evt": "^1.11.2",
     "fastify": "^4.9.2",
     "fastify-metrics": "^9.2.4",
+    "node-pg-migrate": "^6.2.2",
     "p-queue": "^6.6.2",
     "postgres": "^3.3.1",
     "undici": "^5.12.0",
